if customer has an order record can't be delete

+++ eshop/delete_customer.php
@@ -1,26 +1,36 @@
 <?php
 // include database connection
 include 'config/database.php';
-try {     
+try {
     // get record Username
     // isset() is a PHP function used to verify if a value is there or not
-    $Username=isset($_GET['Username']) ? $_GET['Username'] :  die('ERROR: Record Username not found.');
+    $Username = isset($_GET['Username']) ? $_GET['Username'] :  die('ERROR: Record Username not found.');
 
-    // delete query
-    $query = "DELETE FROM customers WHERE Username = ?";
+    $query = "SELECT * FROM orders WHERE customer = ?";
     $stmt = $con->prepare($query);
     $stmt->bindParam(1, $Username);
-     
-    if($stmt->execute()){
-        // redirect to read records page and
-        // tell the user record was deleted
-        header('Location: readCustomers.php?action=deleted');
-    }else{
-        die('Unable to delete record.');
+    $stmt->execute();
+    $row = $stmt->rowCount();
+    if ($row > 0) {
+        header('Location: readCustomers.php?action=notdeleted');
+    } else {
+
+        // delete query
+        $query = "DELETE FROM customers WHERE Username = ?";
+        $stmt = $con->prepare($query);
+        $stmt->bindParam(1, $Username);
+
+        if ($stmt->execute()) {
+            // redirect to read records page and
+            // tell the user record was deleted
+            header('Location: readCustomers.php?action=deleted');
+        } else {
+            die('Unable to delete record.');
+        }
     }
 }
 // show error
-catch(PDOException $exception){
+catch (PDOException $exception) {
     die('ERROR: ' . $exception->getMessage());
 }
 ?>
